Added a function to sign in to audible.com
I don't know for sure if this fetches the right cookies.

Currently we fail with the message "Audible changed their website",
however the error handling and exectution model of GetAllBooks and the
functions it calls is so brittle and cursed that I'm not sure what's
going on. A major overhaul is needed