Initial commit of the rewrite.
We're able to take activation bytes and the path to a HAR file on the
command line, import the cookies from the HAR file, and save that
information in ./.audible-dl.json.  If this file exists, we're able to
read it and dump it to the command line.

Next up I'll be re-integrating the scraper so that we can start
downloading book metadata