Re-integrated the scraper, the HAR file solution works!
We're now able to successfully download a listing of books from
audible using the cookies imported from a HAR file of a request to
audible.com/library/titles while signed in.

Interestingly, while fetching the pages, net/http complains that five
of the cookies contain invalid double quote characters.  Hopefully
we'll be able to remove those cookies and still be able to sign in...