Store partial files in the system's cache directory
This seems like a clearly good solution for partially downloaded
files, however I'm not sure about storing converted files there before
moving them to pwd.  If they're on the same filesystem this isn't a
problem, but my laptop's audiobooks directory is on a different
parition than ~/.cache/, and the os.Rename() operation takes a good
few seconds.

The theory behind converting files into the cache directory is to
avoid littering the audiobooks directory with corrupted files in the
event of premature failure (eg, SIGKILL).  I wonder if there would be
a better way around this; perhaps detaching the ffmpeg processes?