Reworked user interface
This is an ex-nihilo rewrite of the command-line, configuration, and
environment logic for audible-dl that *appears* to replicate the
desired functionality.  I now need to re-work the scraper code for the
new multi-acount functionality and hook them together, then implement
a nice concurrent command-line status.

In the end I'll probably do away with the library/main architecture
and just stick everything in a few package main files.